% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toscana}
\alias{toscana}
\title{A map of Tuscany (Central Italy)}
\format{
\subsection{\code{toscana}}{

An sf object with the boundary of Tuscany in EPSG:32632 crs
sourced from ISTAT files (www.istat.it). It shows the region outline, including the
Tuscan Archipelago, and the provincial borders.
}
}
\usage{
toscana
}
\description{
A map of Tuscany (Central Italy)
}
\keyword{datasets}
