% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toscana_protareas.R
\docType{data}
\name{toscana_protareas}
\alias{toscana_protareas}
\title{A map of Tuscany (Central Italy)}
\format{
\subsection{\code{toscana_protareas}}{

An sf object with the boundary of Tuscan protected areas in EPSG:32632 crs
sourced from Geoscopio files (http://www502.regione.toscana.it/geoscopio/arprot.html).
}
}
\usage{
toscana_protareas
}
\description{
A map of Tuscany (Central Italy)
}
\keyword{datasets}
